+++ cmd/opera/launcher/accountcmd_test.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"io/ioutil"
