+++ cmd/opera/launcher/config.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"bufio"
