+++ cmd/opera/launcher/consolecmd_test.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"crypto/rand"
