+++ cmd/opera/launcher/defaults.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"fmt"
