+++ cmd/opera/launcher/export.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"compress/gzip"
