+++ cmd/opera/launcher/fake.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"crypto/ecdsa"
