+++ cmd/opera/launcher/import.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"bytes"
