+++ cmd/opera/launcher/misccmd.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"fmt"
