+++ cmd/opera/launcher/testdata/dupes/2
