+++ cmd/opera/launcher/testdata/dupes/foo
