+++ cmd/opera/launcher/testdata/empty.js
