+++ cmd/opera/launcher/testdata/guswallet.json
