+++ cmd/opera/launcher/testdata/keystore/.hiddenfile
