+++ cmd/opera/launcher/testdata/keystore/aaa
