+++ cmd/opera/launcher/testdata/keystore/garbage
