+++ cmd/opera/launcher/testdata/keystore/zero
