+++ cmd/opera/launcher/testdata/passwords.txt
