+++ cmd/opera/launcher/testdata/wrong-passwords.txt
