+++ cmd/opera/launcher/validatorcmd.go
@@ -1,4 +1,4 @@
-package main
+package launcher
 
 import (
 	"crypto/ecdsa"
